+++ liara/__init__.py
@@ -32,7 +32,7 @@ from .cache import Cache, FilesystemCache, Sqlite3Cache, RedisCache
 from .util import FilesystemWalker, flatten_dictionary
 from .yaml import load_yaml
 
-__version__ = '2.3.4'
+__version__ = '2.3.5'
 __all__ = [
     'actions',
     'cache',