+++ docs/source/changelog.rst
@@ -1,6 +1,15 @@
 Changelog
 =========
 
+2.3.5 (unreleased)
+------------------
+
+* Improve ``liara validate-links`` command:
+
+  * Check internal links by default. Previously, if run without ``-t``, no links were checked.
+  * Return a non-zero exit code if broken links are found
+  * Add more debug output
+
 2.3.4
 -----